[Net] Fix get_response_body_length for large files.
Parsing was fixed, but not the return value for the exposed getter.

(cherry picked from commit 01e5e98312389f09320f72b34e7be4f5f39f19da)

+++ core/io/http_client.cpp
@@ -611,7 +611,7 @@ Error HTTPClient::poll() {
 	return OK;
 }
 
-int HTTPClient::get_response_body_length() const {
+int64_t HTTPClient::get_response_body_length() const {
 	return body_size;
 }
